Made a test level
I wanted this level to be the first level of the game. It doesn't use the editing mechanic, because I want this level to introduce the players to the game, adding the editing mechanic early on might confuse the players. And, I would like to hear your feedback on this level, send me your feedback in the #ideas-discussion channel.